feat(membershipDiscount): 멤버십 할인 처리
장바구니 항목 Item객체에 영수증 출력을 위해 총구매수량과 증정상품만 있으면 될 줄 알았다
하지만 멤버십할인은 프로모션 미적용 상품에 대해서만 적용된다는 말을 이제서야 깨달았다
다시 일반상품과 프로모션상품 구매수량을 각각 저장하도록 수정한다

+++ src/main/java/store/dto/CalculatePromotionDto.java
@@ -2,6 +2,6 @@ package store.dto;
 
 public record CalculatePromotionDto(
         String name,
-        Integer quantity
+        long quantity
 ) {
 }
